requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.20
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
